Move publish panel to a separate file.
